Replace PortComponent to Excitation
