+++ src/ansys/aedt/core/modeler/circuits/primitives_nexxim.py
@@ -39,7 +39,7 @@ from ansys.aedt.core.internal.load_aedt_file import load_keyword_in_aedt_file
 from ansys.aedt.core.modeler.circuits.object_3d_circuit import CircuitComponent
 from ansys.aedt.core.modeler.circuits.primitives_circuit import CircuitComponents
 from ansys.aedt.core.modeler.circuits.primitives_circuit import ComponentCatalog
-from ansys.aedt.core.modules.boundary.circuit_boundary import Excitations
+from ansys.aedt.core.modeler.circuits.primitives_circuit import Excitations
 
 
 class NexximComponents(CircuitComponents):
